Se pasa metodo del controller para el servicio
